backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
boto3==1.34.69
python-jose==3.3.0
passlib==1.7.4
pydantic
apscheduler==3.10.4
alembic==1.13.1
python-multipart==0.0.9
python-dotenv==1.0.1